Don't build ycm everytime we run update

+++ update.sh
@@ -36,9 +36,18 @@ if [ $? -eq 0 ]; then
     fi
 fi
 
+which yay >/dev/null 2>/dev/null
+if [ $? -eq 0 ]; then
+    yay
+fi
+
 echo 'Updating oh my zsh'
 cd "$HOME/.oh-my-zsh"
 git pull
+
+ycm="$CONF/vimfiles/bundle/YouCompleteMe"
+cd "$ycm"
+OLD_YCM=`git rev-parse HEAD`
 
 echo 'Updating configs'
 cd "$CONF"
@@ -49,11 +58,18 @@ git submodule update --recursive --init
 
 args="--java-completer --clang-completer"
 
+UPD_RUST=0
 which rustup >/dev/null 2>/dev/null
 if [ $? -eq 0 ]; then
+    OLD_RUSTV=`rustc --version`
     echo 'Updating rust'
     rustup update
     args="$args --rust-completer"
+    NEW_RUSTV=`rustc --version`
+
+    if [ "$OLD_RUSTV" != "$NEW_RUSTV" ]; then
+        UPD_RUST=1
+    fi
 fi
 
 which go >/dev/null 2>/dev/null
@@ -72,14 +88,18 @@ if [ $? -eq 0 ]; then
 fi
 
 echo 'Updating vim plugins'
-ycm="$CONF/vimfiles/bundle/YouCompleteMe"
 cd "$ycm"
-python3 ./install.py $args
-cd "$ycm/third_party/ycmd/third_party/racerd"
-git reset . >/dev/null
-git checkout .
-git clean -f -d
-cd "$ycm/third_party/ycmd/third_party/cregex"
-git reset . >/dev/null
-git checkout .
-git clean -f -d
+
+NEW_YCM=`git rev-parse HEAD`
+
+if [ $UPD_RUST -ne 0 -o "$NEW_YCM" != "$OLD_YCM" ]; then 
+    python3 ./install.py $args
+    cd "$ycm/third_party/ycmd/third_party/racerd"
+    git reset . >/dev/null
+    git checkout .
+    git clean -f -d
+    cd "$ycm/third_party/ycmd/third_party/cregex"
+    git reset . >/dev/null
+    git checkout .
+    git clean -f -d
+fi
